+++ src/MonteCarloInterface.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public interface MonteCarloInterface extends Remote {
 
 
-    public int npuntos(int NumeroPares)
+    public int npuntos(long NumeroPares)
             throws java.rmi.RemoteException;
 
 } //end interface
